requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
py3-pinterest==1.4.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.0
requests-toolbelt==1.0.0
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
